Refactor symbologies and enhance parameter handling
- Updated `AvailableSymbologies` enum in `Symbologies.cs` to modify string representations and add new entries.
- Changed `Parameter` in `ValueDouble.cs` to a read-only property; updated `Value` and `Suffix` to be read-only.
- Modified `AvailableParameters` enum in `Parameters.cs` to change device associations and add new parameters for `AvailableDevices.V5`.
- Introduced a new `Custom` class in `Custom.cs` implementing `IParameterValue` with relevant properties and methods.
- Added `CenterPointSortDescription` class in `CenterPointSortDescription.cs` for comparing points.
- Updated `ImageResult.cs` to include primary key attributes for `SourceImageUID` and `ImageRollUID`.
- Modified `LabelVal.csproj` to include a new folder for "Sectors\Messages".
- Changed sorting method in `ImageResultEntry_L95xx.cs` from `SortList` to `SortList2`.
- Refactored `Controller` class in `Controller.cs` for improved operation order and timeout adjustments.
- Enhanced code readability and maintainability across various files.
- Updated `PopupSectorsDetails.xaml` for sorting property name changes.
- Removed static `FallsWithin` method from `ISector` interface, replaced with an extension method.
- Updated `Sector` class to utilize the new `FallsWithin` extension method.
- Improved parameter handling in `SectorDetails` class in `SectorParameters.cs`.
- Updated `ImageResultEntry` classes for V5 and V275 to use the new sorting method and enhance error handling.

+++ ValueConverters/JValueConverter.cs
@@ -22,7 +22,7 @@ namespace JSONViewer_WPF.ValueConverters
                     case JTokenType.Bytes:
                         return (byte[])jval;
                     case JTokenType.Integer:
-                        return jval.Value<Int64>();
+                        return jval.Value<BigInteger>();
                 }
             }
